# Simplify `varray` and `vformat`