@YYF233333 2025-07-02
I changed `vformat`. For `varray` it is both not performace critical and also I think it is more suitable as a macro in its current state (same for `sarray`), as long as you don't mind we have a bit more macros.